Moved tests to type system tests
